refactor(mb-shared): orienter les exports par entity au lieu de la couche api
Remplace les sous-paths `./api/*` (et le baril `./api`) par des exports
directement orientés entity (`./indicateur`, `./referentiel`, `./individu`,
`./me`, `./error`, `./pagination`, `./valeurAvancement`). Les fichiers source
sont déplacés de `src/api/*.ts` vers `src/*.ts` pour aligner la structure
interne sur l'API publique. Tous les consommateurs (mb-api, mb-webapp) sont
mis à jour, y compris les imports multi-entités splittés par entity.
