+++ apps/mb-api/src/indicateur/queries/getIndicateurByPublicId.ts
@@ -1,4 +1,4 @@
-import { type IndicateurApiModel } from '@pilote/mb-shared/api'
+import { type IndicateurApiModel } from '@pilote/mb-shared/indicateur'
 import { ResultAsync } from 'neverthrow'
 
 import { db } from '@/framework/persistence/dbStore'
