+++ apps/mb-api/src/referentiel/routes.ts
@@ -1,13 +1,15 @@
 import { createRoute, OpenAPIHono, z } from '@hono/zod-openapi'
+import { errorApiModelSchema } from '@pilote/mb-shared/error'
 import {
-  createPaginatedApiListSchema,
-  errorApiModelSchema,
   individuApiModelSchema,
   listIndividusForReferentielQuerySchema,
+} from '@pilote/mb-shared/individu'
+import { createPaginatedApiListSchema } from '@pilote/mb-shared/pagination'
+import {
   listReferentielsQuerySchema,
   referentielApiModelSchema,
   referentielPublicIdSchema,
-} from '@pilote/mb-shared/api'
+} from '@pilote/mb-shared/referentiel'
 
 import { never } from '@/framework/errors/never'
 import { jsonResponseOk } from '@/framework/openapi/jsonResponse'
