+++ apps/mb-api/src/valeurAvancement/queries/listIndividusWithValeurs.ts
@@ -1,7 +1,7 @@
 import {
   type IndividusWithValeursListApiModel,
   type ListIndividusWithValeursQuery,
-} from '@pilote/mb-shared/api'
+} from '@pilote/mb-shared/valeurAvancement'
 import { ResultAsync } from 'neverthrow'
 
 import { db } from '@/framework/persistence/dbStore'
