+++ packages/mb-shared/src/error.ts
