+++ packages/mb-shared/src/indicateur.ts
