+++ packages/mb-shared/src/individu.ts
