+++ packages/mb-shared/src/valeurAvancement.ts
